fix: isTransientNetworkError
