+++ src/utils/verifyRecaptcha.ts
@@ -2,28 +2,12 @@ import axios from 'axios';
 import { encode } from 'querystring';
 import config, { isProduction } from '../config';
 import Logger from '../Logger';
+import { isTransientNetworkError, sleep } from './networkRetry';
 
 const GOOGLE_RECAPTCHA_V3_ENDPOINT = config.RECAPTCHA_V3_ENDPOINT;
 
 const MAX_RECAPTCHA_ATTEMPTS = 3;
 const RECAPTCHA_RETRY_BASE_DELAY_MS = 300;
-const TRANSIENT_NETWORK_ERRORS = ['EAI_AGAIN'];
-
-const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
-
-const getNetworkErrorCode = (err: unknown): string | undefined => {
-  if (!axios.isAxiosError(err)) {
-    return undefined;
-  }
-
-  const cause = err.cause as { code?: string } | undefined;
-  return err.code ?? cause?.code;
-};
-
-const isTransientNetworkError = (err: unknown): boolean => {
-  const code = getNetworkErrorCode(err);
-  return !!code && TRANSIENT_NETWORK_ERRORS.includes(code);
-};
 
 async function requestRecaptcha(captcha: string): Promise<boolean> {
   const body = {
